Feat: 시간표 색상 매칭 함수
