+++ src/components/schedule/TimeTable.tsx
@@ -5,32 +5,9 @@ import { Box, Button } from "@mui/material";
 import moment from "moment";
 import { ISchedule } from "../../types/schedule";
 import ScheduleModal from "./ScheduleModal";
+import { createEvents } from "../../features/schedule/createEvents";
 
 const localizer = momentLocalizer(moment);
-
-const createEvents = () => {
-  const today = new Date();
-  const monday = new Date(today.setDate(today.getDate() - today.getDay() + 1));
-
-  const events: ISchedule[] = [
-    {
-      id: "1",
-      title: "김알바",
-      start: new Date(monday.setHours(9, 0, 0)),
-      end: new Date(monday.setHours(10, 30, 0)),
-      color: "#FDA4AF",
-    },
-    {
-      id: "2",
-      title: "이알바",
-      start: new Date(monday.setHours(11, 0, 0)),
-      end: new Date(monday.setHours(12, 30, 0)),
-      color: "#93C5FD",
-    },
-  ];
-
-  return events;
-};
 
 const BoxStyle = {
   height: "800px",
@@ -67,9 +44,6 @@ const BoxStyle = {
   },
   "& .rbc-current-time-indicator": {
     opacity: "0",
-  },
-  "& .rbc-event-selected": {
-    backgroundColor: "transparent",
   },
 };
